Rang Ricochet Nearly Complete
Rang is controlled by movement states. It ricochets off walls until it's been flying for the same amount of time as the first half of a rose petal, at which point it starts accelerating back to the player until it reaches full return speed. When it returns to the player, it goes back to rose mode. There's a bug where the rang returns to rose mode but instead of traveling in the same path it was traveling in before, it bounces forward as if it bounced off of Cotu himself, visually speaking.