Fixed Bounce Bug
When the rang switches from Return to Rose, set_direction is called, which multiplies angle_speed by -1 instead of setting angle_speed to the correct value. Bug fixed by setting angle_speed to the correct value in set_direction (I don't fully understand the values of angle_speed that cause the bug to happen in the moment, but the bug only happens if Cotu was walking to the right when he first threw the rang, meaning the bug was caused by multiplying angle_speed by -1 every time set_direction was called).