Instant Rethrow Saves Buffs
Cotu keeps track of his buffs in a buff list and buffs the Rang when he throws it (if it's an instant rethrow, his buff list should have buffs; if it's not an instant rethrow, his buff list should be empty). Target buffs both Cotu and the Rang. The Rang doesn't keep track of its buffs, it simply updates its attributes (e.g. hitbox damage) whenever it gets buffed.