Minor Rang Collision Improvement
Rang now stays in rose mode and stops colliding with the environment after completing the first half of its current rose petal in rose mode. Before this update, the rang would hit the environment even in the second half of its current rose petal, causing it to switch to return mode. To handle this new functionality, the target now resets the rang's current loop angle to 0 every time the rang hits it.