Fixed Big Enemy Bug
Increased Path Desired Distance, which is the max distance that a nav agent can be from a path point before the path point is considered reached. The nav agent stood still bc it kept trying to go to its next path point, which was directly underneath it. However, since the nav agent was so tall and the PDD was so low, the nav agent kept thinking it had not yet reached its path point, and so it continuously attempted to move straight down onto its path point fruitlessly.