Minor Section Deloading Cleanup
Errors are no longer thrown when a deload is attempted on a deloaded section (by hitting the deload trigger twice). Also, to correct the previous commit's message, there is no need to make death particles and bullets the children of the current section; they can stay children of the level. The only reason to make them children of the section is so that they're deleted along with the rest of the section when the section's deloaded, but if the player doesn't see the section deloading (and the empty space left behind after the section deloads), they won't see the leftover death particles or bullets either, so there's no rush to delete them.