Dodge Run Animation
Created and implemented dodge run animation, which is a third animation in the RunBlendSpace alongside the Idle pose and the normal run animation. Dodge run occurs when the player is moving at the speed of a dodge, but is in the RunBlendSpace instead of a dodge animation.

If Cotu has no vertical motion, Dodge Run is only reachable when Cotu successfully dodges during a crossfade transition from a previous dodge to the RunBlendSpace. If this successful dodge were to happen, an individual animation for it would not play bc animations cannot interrupt Xfade transitions. However, during the Xfade transition, the current RunBlendSpace animation will still play.