New Instant Rethrow CCW Anim
CCW anim now has its own dedicated anim instead of just being CW being played backwards. Why? CW was changed in the last commit so that its keyframes are close together at first, then spread out later on to slow down the spin slightly at the end. If this CW anim were played backwards, it would start slow, then become fast at the end. CCW anim has the same keyframes as CW and they're in the same order, but the keyframes close to the beginning are now spread out and the ones close to the end are close together. This CCW anim is then played backwards in order for it to look almost exactly the same as the CW anim, but counterclockwise.