Jump and Fall Poses
Created Jump and Fall (Ascend and Descend) poses and added them to AnimationTree, which is currently in an unstable state under development. This commit is meant to save progress before I attempt to add the poses to the replacement for RunBlendSpace, which will be MoveBlendSpace, a 2D BlendSpace that includes Idle, Run, DodgeRun, Ascend, and Descend.