Created Miniboss Double Sweep
Also cleaned up enemy code with end_attack func called at the end of every attack, which switches the state back to FOLLOW (previously, a timer in code switched the state). Added a 1 frame await to the beginning of start_attack. Without this await, the animation player would call end_attack at the end of the previous animation on the exact same frame as when the AnimationPlayer.play func is called. Since an animation was currently in progress, the func call would do nothing, leaving the enemy in ATTACK mode but with no animation playing to free it from ATTACK mode, causing it to stand still indefinitely