Fixed MT3 Shield Collision Bug
Rang used to go straight through the shield bc collision.get_collider().collision_layer would return a power of 2 instead of a counting integer in range 1-32. As a result, comparing collision.get_collider().collision_layer to a counting int would always return false if the counting int was any greater than 2, which was true for the ThickEnemy layer but not for the Arena layer, which is why the rang ricocheted correctly when switching the shield to the Arena layer from the ThickEnemy layer.

Fixed the bug by adding a func in the Globals script that compares power of 2 layers (left parameter) to counting int layers (right parameter). The func converts the counting int layer to a power of 2 layer, then compares the values.