Mobile Gunner Animation Polish
Gunner now no longer transitions straight between the WalkBlendSpace and the shoot anim. This is to prevent the gunner from dropping the gun to his side abruptly after every shot (bc it transitions to the Idle anim of the blend space). Gunner now has an aiming animation that plays after every shot. If the gunner is still in attack range after the shot, the aiming anim plays and slowly transitions back to the shoot anim. If the gunner isn't in attack range after the shot, the aiming anim plays briefly and almost immediately transitions to the blend space.