Polished MT1 Anims
Sword is properly between MT1's hands, sword is properly in MT1's hands while walking and idle, bug where sword is at MT1's feet while walking is fixed. The bug came from the lack of SwordPivot position keyframes in the walking anim, which prevented the AnimationTree's BlendSpace from knowing how to interpolate from the Overhead anim's SwordPivot position keyframes to the walking anim's complete lack of position keyframes, so the default of sending the SwordPivot to the position (0,0,0) is done. This bug was fixed by adding SwordPivot position keyframes to the walking anim.