X Boss Triangle Arm Movement
X's arms now go in the correct direction, and triangle_shoot_arms is now properly called in the animation. The bug causing the funcs not to be called was caused by the Triangle animation not having its own file in the filesystem to save the animation data to (I thought it had this already, I must have undone this action in a previous commit). Fixed by reimporting with a proper file