Flying Kick Hitbox Frame Data
Also added a parameter that I manually set to the num of frames that the flying kick hitbox is active (at 60 FPS). This param is fed to the mvmt tween so that X reaches the target in that frame #'s amt of time (the tween takes X from his current position to the target in 9 frames in this case). To do: give superman the same tween-based movement as flying kick