Reworked Buffs
Cotu now has a list of buffs to apply, and when the rang hits the icon, the icon tells Cotu to move up a var indexing into the list so that more buffs in the list are applied. When it's time to apply buffs to the rang, Cotu then looks at each applied buff in the list and does a match case to determine what to do for each buff, which may include funcs that strengthen the rang, icon, or himself.