New Hit and Death Particles
Hit particles are now fluid-like to simulate blood (capsules instead of cubes) and have clear direction because Align-Y is true, which aligns the top of the particle to the direction it's traveling in. Also made the particles last longer, which lets the player see that they're there, and gave the particles gravity to further enforce their existence in the world and display their momentum and fluid-like nature.
Hit particles now match the color of the enemy that got hit.
ALL death particles currently on the floor now match the color of the enemy that most recently just died.
FirstMiniboss now shoots red mega bullets.