Fixed Homing Attack Overshoot Bug
Rang used to overshoot targets due to high speed; fixed by making the rang check if, on the next frame, it would move too far from the target. If so, then the speed would be exactly what it needs to be to get the rang exactly to the target on the next frame. To do: create global trails