Made RightArmSlice look cooler
I attempted to add close range hitboxes (so that the player can't be right next to X and dodge RightArmSlice) by making X kick in front of him, but in every scenario where the close range hitbox lands, so does the blade. As a result, the kicks and other dynamic poses I added just make the attack look a bit cooler and more energetic.