Superman and ChainSlice Damage Rework
Superman will only destabilize if the melee hitbox (his arm) and all explosion hitboxes land (10 melee, 40/30/20 explosion). If all hits of explosion land, Cotu will be left with a sliver of health

If both hits of ChainSlice land, then Cotu dodges, he'll be left with a sliver of health. This allows Cotu to be hit by both hits of ChainSlice and still dodge the following Superman