ChainSlice Visual Tweak
X's arm now follows the blade as it spins around X. Done by attaching an instance of the left arm scene to the blade and toggling its visibility as necessary. Made it so that the laser hitbox on both arms is disabled by default. Also slightly polished positions of blade during ChainSlice and SlipnSlice