TPRightArmSlice
Same as RightArmSlice, but X teleports right on top of the target before doing it (it's actually 1.5m away for optimal slicing). Can only be chosen when X has his right arm, as the move does not recall the arm unlike the original. To do: make X teleport to Cotu's front instead of to his side (assuming Cotu is facing X)