Heavy Ball Complete
Heavy ball now only depresses into the ground upon initial impact when moving at high speeds. If not (e.g. it slides off of another heavy ball already on the ground and gently touches the ground), its entire body stays above the ground and then slowly starts to sink. Its damage is also disabled when it sinks (but not when it slides off of another object)