SemicircleDash Chance Rework
If X does the semicircle dash without his right arm, he'll always do the TPStrafeSlice and get it back (i.e. the anim tree receives the StrafeSlice signal instead of the SemicircleDash signal). If X does the semicircle dash with his right arm, there's a 50% chance he does the TPStrafeSlice and a 50% chance he follows through on the SemicircleDash. Decreased probability that SemicircleDash is chosen when right arm isn't deployed