Fixed Gauntlet Level Lock On Bug
There was a bug where trying to lock on in the Gauntlet would throw an error that said that the cotu var in the Level.gd script was Nil. This didn't make sense since the cotu var was successfully set in the ready func. This bug was fixed by changing the process func in Level.gd to physics process. I still don't know what caused this bug.