Ax Rang Buffs Applied in UI
Ax rang buffs are now visible in the UI (only 1 buff is possible for now). When you perfect catch the ax, an ax rang buff icon appears in the UI under the roserang buff icon. If you catch the ax imperfectly after it's been buffed, the buff disappears. If you perfect catch the ax again after it's been buffed, the icon simply stays.

Unlike with the roserang, the ax applies buffs to the UI using a signal. When the ax instance is deleted, it sends a signal to CotuControl. After receiving the signal, CotuControl checks whether an ax perfect catch input is currently active (queued). If so, a buff is applied in the UI. If a buff was already active in the UI and a perfect catch input isn't currently active, the buff is cleared.

Currently, CotuControl clears the roserang's buffs not when the roserang sends a signal, but as soon as the roserang is seen as null in physics process. This is fine for now because the roserang should immediately lose all buffs as soon as it no longer exists and an instant rethrow didn't just occur (the check for the roserang being null happens after the instant rethrow would make the roserang not null). This paragraph means to say that it's currently unnecessary to make the roserang send a signal when it's deleted to clear its buffs. Null checking in physics process to clear buffs wouldn't work for the ax (without more complications) because it can retain its buffs after being deleted.