Ax Explosion Hitbox and Hit Particle Changes
Added explosion hitbox, explosion particles, and sphere mesh. All of them are controlled via an anim in a new AnimationPlayer.

Hit particles are now emitted from the enemies themselves instead of the thing that hit them.

The process material of the hit particles scene is now Local to Scene (it's a Resource property), which stopped hit particles from all being the same color if multiple enemies were hit by the explosion.