Fixed X Dash Grab Punish Head Rotation Bug
If X grabbed Cotu at the very beginning of the dash, the head would try to look at the real Cotu still on the ground at the start of the punish anim, causing the head to look to the side at a weird angle instead of at the fake Cotu. Fixed by adding a func that makes his animated head node's rotation match his head bone's rotation, then calling it on frame 1 of the anim.