Roserang Homing Instant Rethrow
Roserang can receive a new buff: homing. Upon receiving this, the roserang's instant rethrow becomes a homing throw that autotargets 2 enemies by default (behavior when dodged while it's in rose mode doesn't change; it simply stays in rose mode until instant rethrown). To do: allow the player to dodge the rang when it's returning from a homing instant rethrow (it currently collides with the player no matter what, just like in the special homing throw). To do: make the homing instant rethrow only target enemies in a certain proximity to the rang (when it spawns in) instead of all enemies in the level.

Implementation details:
Add a roserang buff: homing
Replace the first buff in Cotu’s roserang buff list (the damage buff) with the new buff
Create a new enum: roserang_throw_type. This has 2 values: rose (default), homing
Instead of having 2 separate funcs for throwing roserang and throwing special roserang, make 1 func that takes a script as input and sets the roserang’s script to that
In the instant rethrow code in _physics_process, check the throw type. If it’s rose, set roserang script to rose. If it’s homing, set roserang script to homing
Add a new property: homing_targets_added. This is set to the number of times the homing buff appears between indices 0 and next_buff_index-1 in roserang_buff_list (inclusive)
In apply_buffs_to_roserang, if the incoming buff is homing, call roserang_instance.buff_homing_targets(homing_targets_added)
Make a new script: roserang_homing. This is the same as the special_homing script but the max targets starts at 1 instead of 11, and (to do:) only lockonables within a certain radius of the rang are checked (instead of ALL lockonables)
Both roserang_homing and special_homing scripts have a new func: buff_homing_targets(targets_added). This is called in apply_buffs_to_roserang. targets_added is added to the base max_targets property in the homing script, causing the roserang to autotarget more targets.
Why isn’t max_targets simply incremented every time buff_homing_targets is called? Well, what happens when you dodge the rang to give it the homing buff? buff_homing_targets is called, but the rang is still in rose mode, so max_targets in either homing script never increases. We need to store homing_targets_added outside of the homing script because when a homing script is loaded in, max_targets are at their default values instead of the value they should be from the accumulated homing buffs.
To do: make it so that when Cotu dodges the homing rang, it goes back to rose mode