Added Ball Walker Foot Armor
At first, armored plating protects the feet, deflecting the roserang (when it’s not in return mode). This armor has its own health and is destroyed when its HP reaches 0. Armor does not regenerate. Without armor, the feet themselves allow the rose to pass through them

Implementation details:
In a new direct child of Ball Walker (a Node3D), attach a new script called ball_walker_armor.gd. This is mostly the same as EnemyHurtbox, but it’s not a child of Hurtbox nor Area3D (since all it needs to do to receive damage from multihurtbox children is have a receive_hit func). Instead of destroying its parent when its HP reaches 0, it sets its hurtbox children’s process modes (both hurtboxes and physical static bodies) to disabled, and sets the unarmored feet hurtboxes (“flesh” hurtboxes) to inherit (enabled as long as the Ball Walker is enabled)
Why doesn’t the armor destroy itself and its children? So that I can allow the Ball Walker to restore its armor in the future
Create a new StaticBody3D on each standing foot. Each new StaticBody has the ThickEnemy collision layer. A StaticBody is necessary to physically deflect the roserang; an Area3D can’t do it
Create a new multihurtbox child on each standing foot as a sibling of the new StaticBody3D, then set its multihurtbox owner to the ball walker armor