Shrank Flatmite Hurtbox
Flatmite hurtbox is now invulnerable until it jumps due to being too flat. Also here's the logic for the previous commit:

Mvmt states:
Follow:
Neutral: mite sets its target position to a random point near the target, then runs to it. After the leap cooldown ends (can_leap = true), and as soon as the roserang is in flight and far away, the mite enters leap startup mode (in_leap_startup = true)
Leap startup: mite runs directly to the target instead of a random position. Once the mite is close enough, it leaps
Leap: mite leaps into the air and spits at the target. Does a short or long leap depending on how far it is