Mites are Correctly Sized with Slope Walking
Mites used to be just scaled to 1 (instead of .6 and .5 for landmites and paramites, respectively) bc basis was orthonormalized every frame, which sets scale to 1 (along with making vecs orthogonal). Fixed by setting rotation every frame instead of basis; to get rotation, do basis.get_rotation_quaternion for every basis used in the original basis-setting code (and then convert it to a Vector3 afterward using .get_euler() on the lerp result). To do: make flatmite adapt to slopes