Fixed Mite Walking Up Hills Bug
Fixed bug where landmites would slowly walk up hills instead of chasing the target after stepping on certain parts of the terrain. This was fixed by setting path desired distance from 1 to 2 in the landmite nav agent 3D node. With this info, I'm guessing that the bug was caused by the landmite setting its next path position to a point up the hill but underneath the terrain, causing it to move slowly (as most of its movement was in the downward direction). Since the path pos was under the terrain and the path desired dist was too low, the mite couldn't get to it, resulting in it never reaching the point and being able to set another path point. This would also explain the mites following the target again after walking a certain distance up the hill, as they probably reached their desired point finally. I applied this fix (changed path desired dist to 2) for paramites, harvestmen, jumping spider, and flatmite.

Also made landmite tongue hitbox longer to increase accuracy and left the hitbox permanently on the fangs of the landmite when not biting to make landmite more threatening