Save Progress Before Using Coded Raycasts
Will switch to coded raycasts for step targets in harvestman and possibly jumping spider. Node raycasts are turned off when harvestman parent node is deactivated, and for whatever reason, they don't turn on again when parent is reactivated. The following fixes were attempted but the raycasts still didn't turn on:

- From the harvestman parent script, setting RayCast3D nodes’ process modes to ALWAYS and setting their process and physics process states to true
- Calling force_raycast_update and force_update_transform in both process and physics process inside the ray script
- Setting enabled and collide_with_bodies to true every process frame
- Setting ray’s process mode to ALWAYS in the ray script