Harvestman Correctly Spawns from Egg
Fixed bug where harvestman would move under or way above the map when spawned in, which was caused by its feet raycasts not moving to its main body's position before the body's IK starts (I still don't know what exactly the logic is here that causes this to happen). Fixed by adding a short delay in set_active right before the harvestman's process & physics process starts (which lets the harvestman's pos update while it's getting teleported by the egg), and adding another short delay right before the body meshes start IK, which gives the feet IK targets a little time to to move/step into position. The logic is not fully understood at this time, but both of these delays are necessary for the harvestman to function correctly.

To do: try replacing code raycasts in feet with Raycast3D nodes, which should be marginally computationally faster and simpler in implementation