Lingering Roserang Particles
Roserang particles now linger in the air even after the roserang is caught or ricochets. This was done by adding a copy of the particles scene to Cotu, then having Cotu's script move the particles to the rang while it's in flight (Cotu keeps a reference to the current roserang instance). A copy of the rang particles also exists in the roserang scene itself so that the roserang can set the color of the particles based on its current state. Roserang particles are also billboards now instead of flat squares facing upward.