Jumping Spider Jump Dest Calculation Change
Spider now calculates the jump dest as the target + its vel, then moves it slightly closer to itself so it doesn't jump directly on the jump dest, but jumps so that its fangs reach the jump dest. Also gave its eyes a black material so they're not just white. To do: add faraway behaviors (instead of walking, go faraway and summon an egg or leave the arena)