Jumping Spider Fatal Bug Fixed
Walk dest search loop no longer continues indefinitely if no raycast result is found. This was done by setting the condition to an and instead of or, which makes the loop end if either condition is no longer true as intended. Also, the spider only chooses a walk dest within arena_max_radius of the center, preventing it from getting too close to the edge of the arena on a normal walk dest choice