Jumping Spider Faraway State
Jumping spider runs to a random faraway point sometimes instead of walking to a walk dest nearby outside of Cotu's view. Faraway meaning close to the bowl rim, not far from the target. To do: add velocity control so that the spider is more stop and go than high/low velocity