Stop Egg Dropping and Hurtbox Rework
Hurtbox node now receives a reference to its owner (the one whose functions are called and who is queue_freed) instead of assuming its parent is the owner (it doesn't work for jumping spider for example since its hurtbox is a child of its meshes). If no owner is set, parent is used by default. Multi hurtbox owner has not yet been updated. Also mite eggs stop dropping if can_drop is false, and this is set to false when egg waves finish progressing past the final wave